Move refactor code and logic from barrel file
-  This leads to circular dependencies when other code (such as utilities)
   use stuff like ITigerClient etc.

JIRA: RAIL-2999
