+++ libs/api-client-tiger/src/constants.ts
@@ -1,0 +1,7 @@
+// (C) 2019-2021 GoodData Corporation
+export const JSON_API_HEADER_VALUE = "application/vnd.gooddata.api+json";
+
+export const jsonApiHeaders = {
+    Accept: JSON_API_HEADER_VALUE,
+    "Content-Type": JSON_API_HEADER_VALUE,
+};
